Refactor Capcard + security for svgs
