+++ apps/web/app/(org)/dashboard/caps/components/CapCard/CapCardAnalytics.tsx
